// src/generated/models/Cb_todolistsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cb_todolistsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cb_todolistsstatecode = keyof typeof Cb_todolistsstatecode;
export const Cb_todolistsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cb_todolistsstatuscode = keyof typeof Cb_todolistsstatuscode;

export interface Cb_todolistsBase {
  cb_isarchived?: boolean;
  cb_isinbox?: boolean;
  cb_name: string;
  cb_sortorder?: number;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Cb_todolistsstatecode;
  statuscode?: Cb_todolistsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cb_todolists extends Cb_todolistsBase {
  cb_isarchivedname?: string;
  cb_isinboxname?: string;
  cb_todolistid: string;
  createdon?: string;
  modifiedon?: string;
  ownerid: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}